Move write_relay_plugins_toml to a new utils.py module as this is not hermes specific
